Check if hook function exist before unset in case of duplicates
In the case of multiple executables have been sandboxed e.g. via a
multiple shim setup, the code used to try to unset the same hook
multiple times. With this patch, check if the hook is still there before
doing the unset.

I've tested this patch with zsh (5.8) and bash (5.1.8) and it works on
both. Refering to this thread on how to test for function existance in a
way that works for both zsh and bash: https://unix.stackexchange.com/a/332047/19909

Fixes #12